Nixtla TIme Series Predict Model/requirements.txt
Flask==2.2.2
python-dotenv==0.21.0
requests==2.28.1
werkzeug==2.2.3
pandas==2.2.3
nixtla==0.6.6